Add support for systematic tries of FIFO path
This commit adds support to inwee to try different Weechat FIFO paths in
a systematic manner. It first tries the --fp|--fifo-path command line
argument, and then tries the PID FIFO path method, then tries the
default Weechat 1.7 path method- and finally, if all else fails, exits
with an error message.

It introduces a new command line option: --fp | --fifo-path, in which
the user can specify a custom FIFO path for Weechat. In Weechat 1.7 and
later, this is now an option. For backwards compatibility, inwee will
still try the legacy PID method.

Fixes: #1